+++ profiledef.sh
@@ -4,7 +4,7 @@
 iso_name="fwos"
 iso_label="FWOS_$(date --utc --date="@${SOURCE_DATE_EPOCH:-$(date +%s)}" +%Y%m%d)"
 iso_publisher="fwOS project"
-iso_application="fwOS Phase 1 base image"
+iso_application="fwOS Phase 2 basic firewall image"
 iso_version="$(date --utc --date="@${SOURCE_DATE_EPOCH:-$(date +%s)}" +%Y.%m.%d)"
 install_dir="fwos"
 buildmodes=('iso')